Refactoring; More community errors; Qt fix

+++ cnapy/gui_elements/reactions_list.py
@@ -300,7 +300,7 @@ class ReactionList(QWidget):
             self.reaction_mask.is_valid = True
 
             (_, r) = self.splitter.getRange(1)
-            self.splitter.moveSplitter(r/2, 1)
+            self.splitter.moveSplitter(int(r/2), 1)
             self.reaction_list.scrollToItem(item)
             self.reaction_mask.update_state()
